requirements.txt
# Core
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.32.0
httpx>=0.27.0
pydantic>=2.0.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0

# Google APIs (optional - for Gmail/Calendar integration)
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# Scheduling (for 24/7 agent)
schedule>=1.2.0
apscheduler>=3.10.0
